+++ pandas/core/arrays/arrow/__init__.py
@@ -1,3 +1,4 @@
 from pandas.core.arrays.arrow.array import ArrowExtensionArray
+from pandas.core.arrays.arrow.dtype import ArrowDtype
 
-__all__ = ["ArrowExtensionArray"]
+__all__ = ["ArrowDtype", "ArrowExtensionArray"]
